+++ packages/kdot-webdriver/index.js
@@ -1,7 +1,14 @@
 import { merge } from '@generates/merger'
+import { oneLine } from 'common-tags'
 
 export default function kdotWebdriver (config = {}) {
-  const { chrome, firefox, hub } = config
+  const {
+    preset = 'headless',
+    tag = '4.0.0-beta-3-prerelease-20210329',
+    hub,
+    chrome,
+    firefox
+  } = config
   const volumes = [{ name: 'dshm', emptyDir: { medium: 'Memory' } }]
   const volumeMounts = [{ name: 'dshm', mountPath: '/dev/shm' }]
   const shm = { volumeMounts, volumes }
@@ -10,14 +17,22 @@ export default function kdotWebdriver (config = {}) {
     apps: {
       hub: merge(
         {
-          image: {
-            repo: 'selenium/hub',
-            tag: '4.0.0-beta-2-prerelease-20210310'
-          },
+          image: { repo: 'selenium/hub', tag },
           ports: {
-            pub: { port: 4442 },
-            sub: { port: 4443 },
+            pub: { port: 4442, localPort: false },
+            sub: { port: 4443, localPort: false },
             hub: { port: 4444 }
+          },
+          livenessProbe: {
+            httpGet: { path: '/', port: 4444 },
+            initialDelaySeconds: 30,
+            periodSeconds: 5,
+            timeoutSeconds: 1
+          },
+          readinessProbe: {
+            httpGet: { path: '/', port: 4444 },
+            initialDelaySeconds: 15,
+            timeoutSeconds: 1
           }
         },
         hub
@@ -27,19 +42,29 @@ export default function kdotWebdriver (config = {}) {
             chrome: merge(
               {
                 dependsOn: ['hub'],
-                image: {
-                  repo: 'selenium/node-chrome',
-                  tag: '4.0.0-beta-2-prerelease-20210310'
-                },
+                image: { repo: 'selenium/node-chrome', tag },
                 ...shm,
                 env: {
                   SE_EVENT_BUS_HOST: 'hub',
                   SE_EVENT_BUS_PUBLISH_PORT: '4442',
                   SE_EVENT_BUS_SUBSCRIBE_PORT: '4443',
+                  JAVA_OPTS: oneLine`
+                    -Djava.net.preferIPv4Stack=true
+                    -Dwebdriver.chrome.whitelistedIps=
+                  `
+                }
+              },
+              preset === 'debug' && {
+                env: {
                   VNC_NO_PASSWORD: '1'
                 },
                 ports: {
-                  app: { port: 5900, localPort: 5900 }
+                  app: { port: 5900 }
+                }
+              },
+              preset === 'headless' && {
+                env: {
+                  START_XVFB: 'false'
                 }
               },
               chrome
@@ -51,19 +76,26 @@ export default function kdotWebdriver (config = {}) {
             firefox: merge(
               {
                 dependsOn: ['hub'],
-                image: {
-                  repo: 'selenium/node-firefox',
-                  tag: '4.0.0-beta-2-prerelease-20210310'
-                },
+                image: { repo: 'selenium/node-firefox', tag },
                 ...shm,
                 env: {
                   SE_EVENT_BUS_HOST: 'hub',
                   SE_EVENT_BUS_PUBLISH_PORT: '4442',
                   SE_EVENT_BUS_SUBSCRIBE_PORT: '4443',
                   VNC_NO_PASSWORD: '1'
+                }
+              },
+              preset === 'debug' && {
+                env: {
+                  VNC_NO_PASSWORD: '1'
                 },
                 ports: {
                   app: { port: 5900, localPort: 5901 }
+                }
+              },
+              preset === 'headless' && {
+                env: {
+                  START_XVFB: 'false'
                 }
               },
               firefox