+++ packages/kdot/lib/commands/apply.js
@@ -4,10 +4,10 @@ import { kc } from '../k8s.js'
 import emojis from '../emojis.js'
 import getResources from '../getResources.js'
 import applyResource from '../applyResource.js'
-import getRunningPods from '../getRunningPods.js'
+import getReadyPods from '../getReadyPods.js'
 import configure from '../configure/index.js'
 
-const logger = createLogger({ namespace: 'kdot', level: 'info' })
+const logger = createLogger({ namespace: 'kdot.apply', level: 'info' })
 const byTopLevelNamespace = r => !r.app && r.kind === 'Namespace'
 const byTopLevel = r => !r.app && r.kind !== 'Namespace'
 const byDeployment = r => r.kind === 'Deployment'
@@ -26,8 +26,8 @@ export default async function apply (input) {
   const cfg = input.input ? input : await configure(input)
 
   function isRequired (r) {
-    const isArg = r.app && cfg.input.args.includes(r.app.name)
-    return !cfg.input.args.length || isArg || r.app?.isDependency || !r.app
+    const isArg = r.app && cfg.input.args?.includes(r.app.name)
+    return !cfg.input.args?.length || isArg || r.app?.isDependency || !r.app
   }
   function byNew (r) {
     return !r.metadata?.uid && isRequired(r)
@@ -75,11 +75,11 @@ export default async function apply (input) {
   process.stdout.write('\n')
 
   if (cfg.input.wait) {
-    logger.info('Waiting for pods to run...')
+    logger.info('Waiting for pods to be ready...')
     process.stdout.write('\n')
     await Promise.all(resources.filter(byDeployment).map(async deployment => {
       const { namespace, name } = deployment.metadata
-      await getRunningPods(namespace, name, { limit: 1 })
+      await getReadyPods(namespace, name, { limit: 1 })
     }))
   }
 }