+++ packages/kdot/lib/commands/fwd.js
@@ -88,7 +88,8 @@ function forwardPort (app, pod, portConfig) {
  */
 export default async function fwd (input) {
   const cfg = input.input ? input : await configure(input)
-  const apps = Object.values(cfg.apps).filter(a => a.enabled)
+  const bySpecified = input.args?.length ? a => a.isSpecified : a => a.enabled
+  const apps = Object.values(cfg.apps).filter(bySpecified)
   await Promise.all(apps.map(async app => {
     if (app.ports) {
       try {
@@ -96,7 +97,9 @@ export default async function fwd (input) {
         const pod = await getRunningPods(namespace, app.name, pollConfig)
         for (const [name, portConfig] of Object.entries(app.ports)) {
           portConfig.name = name
-          await forwardPort(app, pod, portConfig)
+          if (portConfig.localPort !== false) {
+            await forwardPort(app, pod, portConfig)
+          }
         }
       } catch (err) {
         logger.error(err)