+++ packages/kdot/lib/configure/secrets.js
@@ -2,7 +2,7 @@ import { createLogger } from '@generates/logger'
 import { oneLine } from 'common-tags'
 import encode from '../encode.js'
 
-const logger = createLogger({ namespace: 'kdot.configure', level: 'info' })
+const logger = createLogger({ namespace: 'kdot.cfg.secrets', level: 'info' })
 
 export default function configureSecrets (cfg, owner) {
   // Determine which Secrets to configure and which namespace to use (top-level