+++ packages/kdot/lib/configure/services.js
@@ -1,5 +1,7 @@
 import { merge } from '@generates/merger'
+import { createLogger } from '@generates/logger'
 
+const logger = createLogger({ namespace: 'kdot.cfg.services', level: 'info' })
 const labels = { managedBy: 'kdot' }
 const toServicePorts = (acc, [name, { localPort, reversePort, ...rest }]) => {
   return acc.concat([{ name, ...rest }])
@@ -12,27 +14,28 @@ export default async function configureServices (cfg, owner) {
     const ports = Object.entries(owner.ports).reduce(toServicePorts, [])
     const spec = { selector: { app: name }, ports }
     const metadata = { namespace, name, labels }
-    const service = { kind: 'Service', metadata, spec }
+    const service = { kind: 'Service', app: owner, metadata, spec }
 
     // Merge in any additional service properties specified on the app
     // (e.g. type).
     merge(service, owner.service)
 
     cfg.resources.push(service)
-  } else if (cfg.services) {
+  } else if (!owner && cfg.services) {
     // Configure top-level services.
+    logger.debug('Configure top-level services:', cfg.services)
     for (const [name, given] of Object.entries(cfg.services)) {
       const app = given.app && cfg.apps[given.app]
 
       let ports = given.ports
-      if (!ports && given.app) {
+      if (!ports && app) {
         ports = Object.entries(app.ports).reduce(toServicePorts, [])
       } else {
         throw new Error('No app or port config given for service:', name)
       }
 
       let selector = given.selector
-      if (!selector && given.app) {
+      if (!selector && app) {
         selector = { app: given.app }
       } else {
         throw new Error('No app or selector config given for service:', name)
@@ -40,7 +43,12 @@ export default async function configureServices (cfg, owner) {
 
       const namespace = given.namespace || app.namespace || cfg.namespace
       const metadata = { name, namespace, labels }
-      const service = { kind: 'Service', metadata, spec: { selector, ports } }
+      const service = {
+        app,
+        kind: 'Service',
+        metadata,
+        spec: { selector, ports }
+      }
       cfg.resources.push(service)
     }
   }