+++ packages/kdot-webdriver/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @generates/kdot-webdriver
 
+## 0.1.1
+
+### Patch Changes
+
+- 780e594: Adding JAVA_OPTS settings for chrome
+
+## 0.1.0
+
+### Minor Changes
+
+- af07b9e: Add #150: kdot-webdriver presets
+
+## 0.0.7
+
+### Patch Changes
+
+- 32a1c6b: Specify tag in config for kdot-webdriver
+- 00dbc6b: Adding probes to selenium hub
+
+## 0.0.6
+
+### Patch Changes
+
+- c47991c: Fix #137: Dont forward on localPort: false
+
 ## 0.0.5
 
 ### Patch Changes